# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Procurement Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Register a new vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # RFQs
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth schemas
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        gstNumber:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
        rating:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - companyName
        - contactPerson
        - email
        - phone
        - category
        - status
        - createdAt
    VendorInput:
      type: object
      required: [companyName, contactPerson, email, phone, category]
      properties:
        companyName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        gstNumber:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
    VendorUpdate:
      type: object
      properties:
        companyName:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        gstNumber:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
        rating:
          type: number

    # RFQ schemas
...